controls in slideshow

+++ db/seeds.rb
@@ -10,12 +10,17 @@ class Importer
     Group.create!(:name => 'dinero', :title => 'Dinero', :section => true, :resize_method => 'height', :color => '#FAC9C2')
 
     @images = RAILS_ROOT + "/content/images/"
-    process(@images, load_entries(@images))
+    @posts = RAILS_ROOT + "/content/posts"
+    
+    import_posts(@posts, load_entries(@posts, 'txt'))
+    import_images(@images, load_entries(@images, 'jpg'))
 
     @vacio = Clip.find_by_group_id_and_name(fondos.id, 'vacio')
     Group.all.each do |group|
       group.update_attribute(:background_id, @vacio.id)
     end
+
+    
 
   end
 
@@ -23,22 +28,23 @@ class Importer
     #ActiveRecord::Base.logger = Logger.new(STDOUT)
     Group.destroy_all
     Clip.destroy_all
+    Post.destroy_all
   end
 
   def prepare_directories
     @output = RAILS_ROOT + "/public/miscosillas"
-    FileUtils.remove_dir @output
-    Dir.mkdir(@output) unless File.exist?(@output)
+    FileUtils.remove_dir @output if File.exist?(@output)
+    Dir.mkdir(@output) 
   end
 
 
-  def load_entries(path)
+  def load_entries(path, extension)
     entries = []
-    Dir.chdir(path) { entries = Dir['**/*.jpg'] }
+    Dir.chdir(path) { entries = Dir["**/*.#{extension}"] }
     entries
   end
 
-  def process(base, entries)
+  def import_images(base, entries)
     entries.each do |entry|
       lastBar = entry.rindex('/')
       group = entry[0..(lastBar - 1)]
@@ -53,7 +59,18 @@ class Importer
     end
   end
 
+  def import_posts(base, entries)
+    entries.each do |entry|
+      name = entry
+      path = File.join(base, entry)
+      post = Post.new
+      post.title = name
+      File.open(path, "r") {|f| post.body = f.readlines.join}
+      puts post.save ? "Post #{name} saved!" : "Problem saving post #{name}"
+    end
+  end
 end
+
 
 Importer.new
